Modularize proof of concept
