+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules/
+legacy/
+dist/
 .env
