+++ tsconfig.json
@@ -8,7 +8,7 @@
         "declaration": true,
         "strictNullChecks": true,
         "moduleResolution": "Node16",
-        "target": "ES2020",
+        "target": "ES2021",
         "rootDir": "src",
         "outDir": "lib/cjs",
         "module": "Node16",
